refactor(radio-button): convert files to typescript
