--- src/components/radio-button/index.d.ts
@@ -1,6 +1,0 @@
-export {
-  default,
-  default as RadioButton,
-  PrivateRadioButton,
-} from "./radio-button";
-export { default as RadioButtonGroup } from "./radio-button-group";
