--- src/components/radio-button/index.js
@@ -1,5 +1,0 @@
-import RadioButton, { PrivateRadioButton } from "./radio-button.component";
-import RadioButtonGroup from "./radio-button-group.component";
-
-export { RadioButton, RadioButtonGroup, PrivateRadioButton };
-export default RadioButton;
