+++ src/components/radio-button/index.ts
@@ -1,0 +1,5 @@
+import RadioButton from "./radio-button.component";
+import RadioButtonGroup from "./radio-button-group.component";
+
+export { RadioButton, RadioButtonGroup };
+export default RadioButton;
