+++ src/components/radio-button/radio-button-group.style.ts
@@ -1,6 +1,9 @@
 import styled, { css } from "styled-components";
 
-const RadioButtonGroupStyle = styled.div`
+const RadioButtonGroupStyle = styled.div<{
+  inline?: boolean;
+  legendInline?: boolean;
+}>`
   ${({ inline, legendInline }) => css`
     display: flex;
     flex-direction: column;
