+++ src/components/radio-button/radio-button.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Canvas } from "@storybook/addon-docs";
 import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 
-import { RadioButtonGroup, RadioButton, PrivateRadioButton } from ".";
+import { RadioButtonGroup, RadioButton } from ".";
 import Typography from "../typography";
 
 <Meta title="RadioButton" parameters={{ info: { disable: true } }} />
@@ -554,7 +554,7 @@ Labels could be customized using the Typography Component
 
 ### RadioButton
 
-<StyledSystemProps of={PrivateRadioButton} margin noHeader />
+<StyledSystemProps of={RadioButton} margin noHeader />
 
 **Any other supplied props will be provided to the underlying HTML input element**
